jiffies conversions: Use compile time constants when possible
Do the multiplications and divisions at compile time  instead of runtime when the converted value is a constant.

Make the calculation functions static __always_inline to jiffies.h.

Add #defines with __builtin_constant_p to test and use the static inline or the runtime functions as appropriate.

Prefix the old exported symbols/functions with __